Fix FcrepoEventIT
FcrepoEventIT previously introspected into the contents of the message
body (the JSON message from Fedora).  This poses a problem when
upgrading to fcrepo5, as the message bodies have changed.  Rather than
hard-code the message body assumptions that are true for fcrepo5, the
bodies can be simply ignored, with the bulk of the IT relying on the
message *headers*, as generated by the FcrepoEventProcessor.  The whole
point FcrepoEventProcessor is translating the body into a set of headers
anyway, so this reduced scope is more apt anyway.